Moving _BucketIterator from storage.api to client.
