+++ gcloud/storage/test_api.py
@@ -310,40 +310,6 @@ class Test_create_bucket(unittest2.TestCase):
         self._create_bucket_success_helper('PROJECT', use_default=True)
 
 
-class Test__BucketIterator(unittest2.TestCase):
-
-    def _getTargetClass(self):
-        from gcloud.storage.api import _BucketIterator
-        return _BucketIterator
-
-    def _makeOne(self, *args, **kw):
-        return self._getTargetClass()(*args, **kw)
-
-    def test_ctor(self):
-        connection = object()
-        iterator = self._makeOne(connection)
-        self.assertEqual(iterator.path, '/b')
-        self.assertEqual(iterator.page_number, 0)
-        self.assertEqual(iterator.next_page_token, None)
-
-    def test_get_items_from_response_empty(self):
-        connection = object()
-        iterator = self._makeOne(connection)
-        self.assertEqual(list(iterator.get_items_from_response({})), [])
-
-    def test_get_items_from_response_non_empty(self):
-        from gcloud.storage.bucket import Bucket
-        BLOB_NAME = 'blob-name'
-        response = {'items': [{'name': BLOB_NAME}]}
-        connection = object()
-        iterator = self._makeOne(connection)
-        buckets = list(iterator.get_items_from_response(response))
-        self.assertEqual(len(buckets), 1)
-        bucket = buckets[0]
-        self.assertTrue(isinstance(bucket, Bucket))
-        self.assertEqual(bucket.name, BLOB_NAME)
-
-
 class Http(object):
 
     _called_with = None
